Refactored to make a new CurrentSelection group and moving selection classes there + separating them into several files
